view detail document

+++ app/Http/Controllers/User/DocumentController.php
@@ -3,27 +3,38 @@
 namespace App\Http\Controllers\User;
 
 use Auth;
-// use Exception;
+use Storage;
+use Session;
+use Exception;
 use Illuminate\Http\Request;
 use App\Traits\UploadFileTrait;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UploadDocumentRequest;
+use App\Repositories\Contracts\UserRepositoryInterface;
+use App\Repositories\Contracts\BookmarkRepositoryInterface;
 use App\Repositories\Contracts\CategoryRepositoryInterface;
 use App\Repositories\Contracts\DocumentRepositoryInterface;
+
 
 class DocumentController extends Controller
 {
     use UploadFileTrait;
 
+    protected $userRepository;
     protected $categoryRepository;
     protected $documentRepository;
+    protected $bookmarkRepository;
 
     public function __construct(
+        UserRepositoryInterface $userRepository,
         CategoryRepositoryInterface $categoryRepository,
-        DocumentRepositoryInterface $documentRepository
+        DocumentRepositoryInterface $documentRepository,
+        BookmarkRepositoryInterface  $bookmarkRepository
     ) {
+        $this->userRepository = $userRepository;
         $this->categoryRepository = $categoryRepository;
         $this->documentRepository = $documentRepository;
+        $this->bookmarkRepository = $bookmarkRepository;
     }
 
     /**
@@ -91,10 +102,38 @@ class DocumentController extends Controller
      */
     public function show($id)
     {
-        $document = $this->documentRepository->find($id);
-        $relatedDocuments = $this->documentRepository->where('category_id', '=',$document->category_id)->with('user')->get()->take(10);
+        try {
+            $document = $this->documentRepository->getDocument($id);
 
-        return view('user.pages.view-document', compact('document', 'relatedDocuments'));
+            if (Session::has('recently_view')) {
+                $recentlyView = Session::get('recently_view');
+                if (!in_array($id, $recentlyView)) {
+                    $views = $document->views + 1;
+                    $this->documentRepository->update($id, ['views' => $views]);
+                    Session::push('recently_view', $document->id);
+                }
+            } else {
+                $views = $document->views + 1;
+                $this->documentRepository->update($id, ['views' => $views]);
+                Session::push('recently_view', $document->id);
+            }
+
+            $relatedDocuments = $this->documentRepository->where('category_id', $document->category_id)
+                ->where('id', '!=', $document->id)
+                ->with('user')->get()->take(10);
+            $authorUploaded = $this->documentRepository->where('status',config('settings.document.status.is_published'))
+                ->where('user_id', $document->user->id)->count();
+
+            if (Auth::check()) {
+                $isBookmark = $this->bookmarkRepository->isBookmark(Auth::user()->id, $id);
+            } else {
+                $isBookmark = config('settings.document.is_bookmark.false');
+            }
+
+            return view('user.pages.view-document', compact('document', 'relatedDocuments', 'authorUploaded', 'isBookmark'));
+        } catch(Exception $e) {
+            return back()->with('messageError', trans('user.document.document_not_found'));
+        }
     }
 
     /**
@@ -105,9 +144,17 @@ class DocumentController extends Controller
      */
     public function edit($id)
     {
-        $document = $this->documentRepository->find($id);
+        try {
+            $document = $this->documentRepository->findOrFail($id);
 
-        return view('user.pages.edit-document', compact('document'));
+            if ($document->user_id == Auth::user()->id) {
+                return view('user.pages.edit-document', compact('document'));
+            }
+
+            return back();
+        } catch(Excepion $e) {
+            return back();
+        }
     }
 
     /**
@@ -131,9 +178,62 @@ class DocumentController extends Controller
     public function destroy($id)
     {
         try {
-            $this->documentRepository->delete($id);
+            $document = $this->documentRepository->findOrFail($id);
 
+            if ($document->user_id == Auth::user()->id) {
+                $this->documentRepository->delete($id);
+
+                return back()->with('messageSuccess', trans('user.document.delete_success'));
+            }
+
+            return back()->with('messagesError', trans('user.document.delete_fail'));
+        } catch(Exception $e) {
+            return back()->with('messagesError', trans('user.document.delete_fail'));
+        }
+    }
+
+    public function download($id)
+    {
+        try {
+            $document = $this->documentRepository->find($id);
+            $fileName = $document->file_name;
+            $filePath = str_replace('storage', '', $fileName);
+            $exist = Storage::disk('public')->exists($filePath);
+
+            if ($exist) {
+                $downloads = $document->downloads + 1;
+                $this->documentRepository->update($id, ['downloads' => $downloads]);
+
+                if (Auth::check()) {
+                    $downloadedDocument = Auth::user()->downloaded;
+                    $downloadedDocument = explode(',', $downloadedDocument);
+                    
+                    if (!in_array($id, $downloadedDocument)) {
+                        array_push($downloadedDocument, $id);
+                    }
+
+                    $downloadedDocument = implode(',', $downloadedDocument);
+                    $this->userRepository->update(Auth::user()->id, ['downloaded' => $downloadedDocument]);
+                }
+
+                return response()->download($document->download_link);
+            } else {
+                return back();
+            }
+        } catch(Exception $e) {
             return back();
+        }
+    }
+
+    public function showDownloaded()
+    {
+        try {
+            $user = Auth::user();
+            $downloadedDocument = explode(',', $user->downloaded);
+            $downloadedDocuments = $this->documentRepository->withTrashed()->whereIn('id', $downloadedDocument)
+                ->paginate(config('settings.document.downloaded.paginate'));
+
+            return view('user.pages.downloaded', compact('downloadedDocuments'));
         } catch(Exception $e) {
             return back();
         }
